connect board crud

+++ client/src/components/pages/dashboards/use-delete.ts
@@ -2,19 +2,18 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { boardApi } from './api';
 
-export const useDeleteBoard = () => {
+export const useDeleteBoard = (userId: string) => {
   const queryClient = useQueryClient();
 
   const deleteMutation = useMutation({
     mutationFn: (id: string) => {
       const abortController = new AbortController();
 
-      return boardApi.deleteBoard(id)({ signal: abortController.signal });
+      return boardApi.deleteBoard(id, userId)({ signal: abortController.signal });
     },
   });
 
   const onDelete = (id: string) => {
-    console.log({ deleteId: id });
     deleteMutation.mutate(id, {
       onSuccess: () => {
         queryClient.invalidateQueries({
